[ui] move variants and hook to separate files
